requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.6.0
pydantic-settings==2.4.0
python-multipart==0.0.9
